fix: stabilize inline rename input and e2e locators
Component:
- Add data-testid/data-tab-title to the outer tab div so tests can target
  tabs by title reliably, including while the title is inside an input.
- Stop propagation on the rename input's onAuxClick and preventDefault on
  middle-click mouseDown so the Linux X11 primary-selection paste is
  blocked inside the rename field too, matching the outer tab behavior.
- Document why handleRenameOpen snapshots the title once and why the focus
  effect is scoped to [isEditing] — neither should react to external title
  changes mid-edit, or the user's typing could be overwritten/refocused.

Tests:
- Move inline-rename specs to tab-rename.spec.ts (tabs.spec.ts was over
  oxlint's 300-line cap).
- tabLocatorByTitle uses a CSS-escaped data-attr locator, avoiding the
  unanchored/unescaped-regex bug in getByText/getByRole regex locators.
- Use { name, exact: true } on the rename input locator.
- beforeEach clears leftover customTitle so tests sharing an Electron
  worker see default titles, not prior renames.
- Switch .toBeNull() to .toBe(null) under expect.poll and poll the Escape
  assertion to remove a race.
- After Enter commits, also assert the new title is visible in the DOM.
