+++ tests/e2e/tab-rename.spec.ts
@@ -1,0 +1,170 @@
+/**
+ * E2E tests for inline tab renaming (double-click a tab to rename).
+ *
+ * User Prompt:
+ * - double-click a tab to rename it inline
+ */
+
+import { test, expect } from './helpers/orca-app'
+import {
+  waitForSessionReady,
+  waitForActiveWorktree,
+  getActiveWorktreeId,
+  getActiveTabId,
+  getWorktreeTabs,
+  ensureTerminalVisible
+} from './helpers/store'
+
+test.describe('Tab Rename (Inline)', () => {
+  test.beforeEach(async ({ orcaPage }) => {
+    await waitForSessionReady(orcaPage)
+    await waitForActiveWorktree(orcaPage)
+    await ensureTerminalVisible(orcaPage)
+    // Why: clear any custom titles left by a previous test (the Electron app
+    // persists across tests in the worker) so tab locators key off the default
+    // title, not a stale rename like "My Custom Title".
+    await orcaPage.evaluate(() => {
+      const store = window.__store
+      if (!store) {
+        return
+      }
+      const state = store.getState()
+      for (const tabs of Object.values(state.tabsByWorktree)) {
+        for (const tab of tabs) {
+          if (tab.customTitle != null) {
+            state.setTabCustomTitle(tab.id, null)
+          }
+        }
+      }
+    })
+  })
+
+  async function getActiveTabTitle(
+    page: Parameters<typeof getActiveTabId>[0],
+    worktreeId: string
+  ): Promise<string> {
+    const activeId = await getActiveTabId(page)
+    expect(activeId).not.toBeNull()
+    const tabs = await getWorktreeTabs(page, worktreeId)
+    const tab = tabs.find((entry) => entry.id === activeId)
+    expect(tab).toBeDefined()
+    // Why: mirror what the UI renders (customTitle ?? title) so locators that
+    // key off the tab's visible text match what's actually on screen.
+    return tab!.title ?? ''
+  }
+
+  function tabLocatorByTitle(
+    page: Parameters<typeof getActiveTabId>[0],
+    title: string
+  ): ReturnType<Parameters<typeof getActiveTabId>[0]['locator']> {
+    // Why: backslash first so the backslashes we introduce when escaping the
+    // double-quote aren't themselves re-escaped; both chars are CSS-selector
+    // metacharacters inside a double-quoted attribute value.
+    const escaped = title.replace(/\\/g, '\\\\').replace(/"/g, '\\"')
+    return page.locator(`[data-testid="sortable-tab"][data-tab-title="${escaped}"]`).first()
+  }
+
+  async function getActiveCustomTitle(
+    page: Parameters<typeof getActiveTabId>[0],
+    worktreeId: string
+  ): Promise<string | null> {
+    return page.evaluate((targetWorktreeId) => {
+      const store = window.__store
+      if (!store) {
+        return null
+      }
+
+      const state = store.getState()
+      const activeId = state.activeTabIdByWorktree[targetWorktreeId] ?? state.activeTabId
+      const tab = (state.tabsByWorktree[targetWorktreeId] ?? []).find((t) => t.id === activeId)
+      return tab?.customTitle ?? null
+    }, worktreeId)
+  }
+
+  test('double-clicking a tab opens an inline rename input and Enter commits', async ({
+    orcaPage
+  }) => {
+    const worktreeId = (await getActiveWorktreeId(orcaPage))!
+    const originalTitle = await getActiveTabTitle(orcaPage, worktreeId)
+    expect(originalTitle.length).toBeGreaterThan(0)
+
+    const tabLocator = tabLocatorByTitle(orcaPage, originalTitle)
+    await tabLocator.dblclick()
+
+    const renameInput = orcaPage.getByRole('textbox', {
+      name: `Rename tab ${originalTitle}`,
+      exact: true
+    })
+    await expect(renameInput).toBeVisible()
+
+    await renameInput.fill('My Custom Title')
+    await renameInput.press('Enter')
+
+    await expect
+      .poll(async () => getActiveCustomTitle(orcaPage, worktreeId), { timeout: 3_000 })
+      .toBe('My Custom Title')
+    await expect(renameInput).toBeHidden()
+    await expect(tabLocatorByTitle(orcaPage, 'My Custom Title')).toBeVisible()
+  })
+
+  test('Escape during inline rename discards the edit', async ({ orcaPage }) => {
+    const worktreeId = (await getActiveWorktreeId(orcaPage))!
+    const originalTitle = await getActiveTabTitle(orcaPage, worktreeId)
+
+    const tabLocator = tabLocatorByTitle(orcaPage, originalTitle)
+    await tabLocator.dblclick()
+
+    const renameInput = orcaPage.getByRole('textbox', {
+      name: `Rename tab ${originalTitle}`,
+      exact: true
+    })
+    await expect(renameInput).toBeVisible()
+
+    await renameInput.fill('Should Be Discarded')
+    await renameInput.press('Escape')
+
+    await expect(renameInput).toBeHidden()
+    await expect
+      .poll(async () => getActiveCustomTitle(orcaPage, worktreeId), { timeout: 3_000 })
+      .toBe(null)
+  })
+
+  test('renaming to an empty string resets the tab to its default title', async ({ orcaPage }) => {
+    const worktreeId = (await getActiveWorktreeId(orcaPage))!
+
+    // Why: seed a custom title directly via the store so this test asserts the
+    // "empty string → reset" behavior independently from the double-click flow.
+    await orcaPage.evaluate((targetWorktreeId) => {
+      const store = window.__store
+      if (!store) {
+        return
+      }
+
+      const state = store.getState()
+      const activeId = state.activeTabIdByWorktree[targetWorktreeId] ?? state.activeTabId
+      if (activeId) {
+        state.setTabCustomTitle(activeId, 'Seeded Custom')
+      }
+    }, worktreeId)
+
+    await expect
+      .poll(async () => getActiveCustomTitle(orcaPage, worktreeId), { timeout: 3_000 })
+      .toBe('Seeded Custom')
+
+    const tabLocator = tabLocatorByTitle(orcaPage, 'Seeded Custom')
+    await tabLocator.dblclick()
+
+    const renameInput = orcaPage.getByRole('textbox', {
+      name: 'Rename tab Seeded Custom',
+      exact: true
+    })
+    await expect(renameInput).toBeVisible()
+
+    await renameInput.fill('')
+    await renameInput.press('Enter')
+
+    await expect
+      .poll(async () => getActiveCustomTitle(orcaPage, worktreeId), { timeout: 3_000 })
+      .toBe(null)
+  })
+})
